Add suport to "data-width" in <span> elements
According https://github.com/scottjehl/picturefill/issues/87 issue, some
people miss the feature of specify width of generated image. This pull
request add the capability of do this by specifying a `data-with` attr
in the respective `<span>`s which are supposed to have this.

The README.md wasn't updated because I don't knew in what places to put
the infos to not to lose the "line" of the explanations.