Check if picImg is already a child of matchedEl
In working on some local modifications to picturefill (which I hope to submit as a PR at a later time), I noticed some jerkiness when resizing the page. I found that the end action where the code changes the `src` and re-appends the element were the source of the problem.

Since these actions are unnecessary if we are not changing the image, this adds a `continue` if the image's parent node is already `matchedEl`.