Added option to disable automated defining in complicated environment
When picturefill needs to be loaded in an environment which also have a package loader active (which is loaded further down the page) it's handy to be able to load picturefill async without registering it as a module.

This commit takes care of that. You can disable auto definition like this:
```
<script>var picturefillConfig = {define: false};</script>
<script src="js/vendor/polyfill/picturefill.min.js")" async></script>
```