Adjusts image width/height based on candidate resolution
Previously, if a candidate was selected with a resolution > 1, no
attempt was made to adjust this, so that the image would display at the
correct size.

This, once the image candidate is loaded, get the natural width and
height of the image (not supported in IE<=8), and divide that by the
image resolution.

By using html width/height attributes rather than CSS width/height, it
allows the dimensions to be easily overwritten in user CSS without
requiring an !important declaration.